# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: API for the WASAText messaging application.
  version: "1.0.2"
servers:
  - url: http://localhost:3000 

#logical endpoints grouping for better documentation and client generation
...
security:
  - bearerAuth: []

#endpoints
paths:
...
  /groups/{groupId}/members:
    post:
      tags: [groups]
      summary: Add a user to a group
      description: Add a user to a group he is not in
      operationId: addToGroup
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToGroupRequest"
      responses:
        "200":
          description: User added to group successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Invalid request or user is not in the group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already belongs to the group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  #endpoint that allows tha authenticated user to leave a group conversation they are a member of
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UserIdentifier

  #standard input parameters that are reused across multiple endpoints to avoid repetition and ensure consistency in parameter definitions
  parameters:
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group conversation identifier
      schema:
        type: string
        minLength: 4
        maxLength: 64
        pattern: '^.+$'
...
  schemas:
...
    photo:
      type: string
      minLength: 1
      maxLength: 512
      description: URL or local path of the profile/group photo
      example: "/static/users/user123.png"
      pattern: '^(https?://.+|/.+)$'

    #schema that creates an object for creating a new conversation, either direct or group.
    #for direct conversations, only one member (other than the creator) should be specified. For group conversations, at least one member should be specified and a name is required
...
    AddToGroupRequest:
      type: object
      description: object for adding a user to a group conversation
      required:
        - userId
      properties:
        userId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the user to be added to the group chat
          example: "user_001"

    #schema that creates an object for updating the name of a group conversation
...
    User:
      type: object
      description: user profile
      required:
        - id
        - username
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
          description: the user id
          pattern: '^.*?$'
          example: "user_001"
        username:
          type: string
          minLength: 1
          maxLength: 64
          description: the username
          pattern: '^.*?$'
          example: "Maria_1604"
        photo:
          $ref: "#/components/schemas/photo"

    #schema representing a list of users, used for responses that return multiple user profiles
...
    Conversation:
      type: object
      description: conversation with all its messages
      required:
        - id
        - isGroup
        - members
        - messages
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the conversation
          example: "conv_001"
        isGroup:
          type: boolean
          description: true if the conversation is a group conversation
          example: true
        name:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the name of the conversation
          example: "Study Group"
        photo:
          $ref: "#/components/schemas/photo"
        members:
          type: array
          minItems: 0
          maxItems: 16
          description: an array of users members of the conversation
          items:
            $ref: "#/components/schemas/User"
        messages:
          type: array
          description: an array of messages sent in the conversation
          minItems: 0
          maxItems: 4096
          items:
            $ref: "#/components/schemas/Message"

    #schema representing a message in a conversation, which can be of type text, gif, or image
    #the content field represents the message content, which is either the text itself for text messages or a URL string for media messages
    #the message also includes metadata such as the sender, creation time, delivery and read status, and any comments/reactions associated with it
    Message:
      type: object
      description: message in a conversation
      required:
        - id
        - conversationId
        - senderId
        - type
        - content
        - createdAt
        - status
        - comments
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the message
          example: "msg_001"
        conversationId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the conversation the message is in
          example: "conv_001"
        senderId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the user id of the sender of the message
          example: "user_001"
        senderUsername:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the username of the sender of the message
          example: "Maria-0327"
        type:
          type: string
          minLength: 3
          maxLength: 5
          pattern: '[a-z]'
          description: the type of the message (text, gif, image)
          enum: [text, gif, image]
          example: text
        content:
          type: string
          minLength: 1
          maxLength: 4096
          pattern: '^.*?$'
          description: the content of the message
          example: "Hello everyone"
        replyToMessageId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the message that this message is replying to. Null if not replying to any message
          example: "msg_000"
        forwardedFromMessageId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the message that this message is being forwarded from. Null if not forwarded from any message
          example: "msg_777"
        createdAt:
          type: string
          minLength: 1
          maxLength: 32
          pattern: '^.*?$'
          description: timestamp for when this message has been sent
          format: date-time
        deleted:
          type: boolean
          description: true if this message has been deleted from the sender
          example: false
        status:
          $ref: "#/components/schemas/MessageStatus"
        comments:
          type: array
          description: an array of comments that are attached to this message
          minItems: 0
          maxItems: 16
          items:
            $ref: "#/components/schemas/Comment"

    #schema representing the delivery and read status of a message, indicating whether it has been delivered to all conversation members and whether it has been read by all members
    MessageStatus:
      type: object
      description: delivery and read status of a message, indicating whether it has been delivered to all conversation members and whether it has been read by all members
      required:
        - deliveredToAll
        - readByAll
      properties:
        deliveredToAll:
          description: true if delivered by all the conversation members
          type: boolean
          example: true
        readByAll:
          type: boolean
          description: true if read by all the conversation members
          example: false

    #schema representing a comment added to a message, which includes the comment content, the author, and the creation time
    Comment:
      type: object
      description: comment added to a message
      required:
        - id
        - messageId
        - authorId
        - content
        - createdAt
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the comment
          example: "comm_001"
        messageId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the message this comment is attached to
          example: "msg_001"
        authorId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the id of the user that attached this comment
          example: "user_001"
        authorUsername:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*?$'
          description: the username of the user that attached this comment
          example: "Luca"
        content:
          type: string
          minLength: 1
          maxLength: 32
          pattern: '^.*?$'
          description: the comment (Emoticon)
          example: "😀"
        createdAt:
          type: string
          minLength: 1
          maxLength: 32
          pattern: '^.*?$'
          description: the timestamp at wich this comment was added
          format: date-time
    Error:
      type: object
      description: an error object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4096
          pattern: '^.*?$'
          description: the string that explains the error
          example: "invalid request"